working on separating vehicle details and edit forms' presentation from logic
